+++ packages/streamdeck-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fcannizzaro/streamdeck-react
 
+## 0.2.5
+
+### Patch Changes
+
+- bump versions, update samples, add versions manifest for native modules
+
 ## 0.2.4
 
 ### Patch Changes